Added `-sink-insts-to-avoid-spills` llc flag to workflow.
In simple loops where we need to call a function and pass by reference,
the X86 backend uses a LEA instruction to get the reference, which is hoisted
outside of the loop. This causes greater register pressure and might
lead to a spill, that AArch64 will not have.

Thus, we add the llc flag to sink the LEA instruction back into the
loop, since it causes greater register pressure. Maybe a temporary fix.

Closes #30