Added build functionality to the NPB Runner
This flag-invoked functionality builds the executable list specified in
the config by jumping to NPB_PATH and invoking the Makefile. We assume
for now that the build directory for each benchmark can be extracted
from the first letters of the exe name. E.g., for
is_aarch64_aligned.out, we will build inside the is/ directory, located
in NPB_PATH.