Add post-process capability to the NPB runner
Currently, this capability is run on the output folder of the
experiment. Rougly speaking, it extracts the execution times from the
run results, it averages them, and it creates a csv with all the
benchmarks per experiment.

Also, this commit adds the respective post-process scripts,
to extract csvs from experiments.