Fix small stackmap dump bug.
When giving the callsite to dump as an argument,
the id was taken as a global callsite id, not local to the function.
This commit fixes that so when you give a specific callsite id, this
will be counted from the parent function, not the whole callsite/record
collection in the stackmap (which are stored sequentially for all
functions of the module).