Add option to sort data in plot_barchart
The index of the DataFrame that is plotted is usually the name of the
benchmarks. It's useful to have the option to sort them alphabetically,
so that different unordered data sheets can be plotted in the same fashion.
Otherwise, we have to impose an order ourselves in the data sheet.

This is added as an option in case we actually want to show a different
order than the alphabetical.