Sort benchmarks lexicographically (lowercase first) when plotting
When we are plotting benchmarks it is usually desired to have some
aggregate numbers in the end. If we sort the index, these numbers will
be mixed with the benchmarks. A compromise here is to assume that the
benchmarks are lowercase and the aggregate numbers are uppercase. This
way, we can sort them in a clean way lexicographically.