Add separate flags for unmodified LLVM and musl toolchains
When doing the breakdown experiments, sometimes we may need to have a
(partially) modified LLVM toolchain, but the unmodified musl toolchain.
For example, we may want to keep only the aligment changes, but not the
register file changes. In that case, we need to use a modified LLVM
toolchain, but an unmodified musl toolchain.

Unmodified LLVM: UNMODIFIED_LLVM
Unmodified musl: UNMODIFIED_MUSL
Both: UNMODIFIED (keeps the previous behaviour)