[layout] Refactor common.mk llc-related flags
Split flags into categories, aligning them with the LLVM modification categories
described in our CGO 2024 submission. This makes experimenting with the
performance breakdown easier (can detect which flags to keep or not),
and gives better overview of the flags.

Adds a bit more comments to this Makefile section as well.