[layout] Add flag to avoid inserting stackmaps during compilation
Sometimes, compilation fails when lowering the stackmaps, because of
a potential unhandled case. We add the `NO_STACKMAPS` compilation option
in the common layout Makefile, so that we can skip the stackmap
insertion. This is useful when we are breaking down the overhead of some
modifications, and we want to skip temporarily stackmaps bugs.