[layout] Add the `deepclean` target to the common.mk
Now, the `clean` target will not delete the build directories, only the
intermediate compilation files. Only with the `deepclean` flag, will the
build directories be deleted as well.

This is helpful if we want to keep the debug files inside the build
dirs, which come from commands like `-print-before`/`-print-after` pass
in llc.