[layout] Add stack slot align utilities (#317)
Similar to callsite align, adds scripts to align stack slots between arm
and x86 to the maximum alignment. In particular, it takes as input the
output for both architectures of the `stack-slot-coloring` pass, parses
for each function all the stack slots and their alignment, and dumps a
json file with the function name and the stack slots along with the max
alignment between the architectures per stack slot.

Also, adds utility scripts to check the alignment. These scripts also
detect different number of functions/stack slots. They only exit if the
number of functions is different and print a warning in the case of
different slots.