[Intel Pin Runner] Add support for mangled function names
Since we want to be able to list all function symbols, even if they are
mangled, we extend the `nm` output preprocessing.

First, we do not filter out functions starting with '_'. This will be
checked inside the Pin Tool.

Then, we remove trailing `.[....]` strings that sometimes occur with openmp
(e.g., `.[fomp...]`.